Create generic survey form component
